requirements.txt
python-chess>=1.9.0
numpy>=1.21.0
pytest>=7.0.0
